networking: assign and release public IP implementation
The methods assignPubIP and releasePubIP were missing their implementation.
In order to implement these functions these code relies the corresponding
iptables chains are in the cnci agent.

To assign public IP:
  - Assign the public ip to the iface of the CNCI Agent
  - For incoming traffic we are forwarding of all traffic coming to the
public IP assigned to the cnci node to the private IP address of instance.
we do this by making a DNAT. (ciao-floating-ip-pre chain)
  - For outgoing traffic we are masking the traffic by implementing a nat
rule for SNAT, so the outgoing traffic from the instance will go to the
cnci node and then to the destitation as if it were coming directly from
the instance. (ciao-floating-ip-post chain)
  - Every time we add a new rule, we insert it at the beggining of its
corresponding defined chain.

To release public IP:
  - Remove the public ip from the iface of the CNCI Agent
  - Verify the existing rule for the given IPs (internal and external)
exist in each chain (ciao-floating-ip-pre and ciao-floating-ip-post),
and then delete them, if they don't exist then we just notify.

Fixes #532

Signed-off-by: Leoswaldo Macias <leoswaldo.macias@intel.com>