ciao-controller: include CNCI UUID in AssignPublicIP payload
The controller knows internally which CNCI needs to handle a floating
IP assignment, but the ssntp server aka ciao-scheduler does not.
This information must be included in the payload for the scheduler to
send the command to the correct CNCI.

This requires accessing a datum from the controller's internal datastore.

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>
Signed-off-by: Leoswaldo Macias <leoswaldo.macias@intel.com>