ciao-scheduler: tidy some comments/logs/naming in CNCI event forwarding
After looking at this code to make a similar flow for CNCI command
forwarding it was clear it needed some tidying up.  The function name was
not specific enough.  The comment wasn't clear enough.  The logged messages
had typos, likely printed a nul value, and weren't clear enough to help a
reader of the log know what they're seeing without comparing to the source.

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>